Move test code to the right place
